# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ghost Shift HTTP API
  version: 0.1.0
  description: Public, internal, operational, and observability endpoints exposed by Ghost Shift.
servers:
  - url: /
paths:
...
  /api/public/timeline:
    get:
      summary: Get public timeline points
      operationId: getPublicTimeline
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Timeline response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTimelineResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/public/replay:
    get:
      summary: Get public replay frames
      operationId: getPublicReplay
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Replay response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicReplayResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/public/analytics/trends:
    get:
      summary: Get trend changes across the past N hours
      operationId: getAnalyticsTrends
      parameters:
        - $ref: "#/components/parameters/Hours"
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Trend response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTrendResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/public/zones/heatmap:
    get:
      summary: Get zone activity heatmap
      operationId: getZonesHeatmap
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Zone heatmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicZonesHeatmapResponse"
  /api/public/models/distribution:
    get:
      summary: Get model distribution metrics
      operationId: getModelsDistribution
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Model distribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicModelsDistributionResponse"
...
  /internal-api/presence:
    get:
      summary: Get gateway presence entries
      operationId: getInternalPresence
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Presence list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PresenceEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /internal-api/channels:
    get:
      summary: Get gateway channels
      operationId: getInternalChannels
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Channel list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /internal-api/nodes:
    get:
      summary: Get gateway nodes
      operationId: getInternalNodes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Node list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NodeInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /internal-api/cron:
    get:
      summary: Get gateway cron jobs
      operationId: getInternalCron
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cron list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CronJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    Hours:
      name: hours
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    Since:
      name: since
      in: query
      required: false
      schema:
        type: string
        format: date-time
    Until:
      name: until
      in: query
      required: false
      schema:
        type: string
        format: date-time
  responses:
    BadRequest:
      description: Invalid query parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid internal API token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        requestId:
          type: string
      required: [error]
...
    PublicStatusShare:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
        share:
          type: number
      required: [status, count, share]
    PublicZoneMixEntry:
      type: object
      properties:
        zone:
          type: string
        count:
          type: integer
        share:
          type: number
      required: [zone, count, share]
    PublicModelMixEntry:
      type: object
      properties:
        model:
          type: string
        count:
          type: integer
        share:
          type: number
      required: [model, count, share]
    PublicSession:
      type: object
      properties:
        publicId:
          type: string
        sessionKey:
          type: string
        agentId:
          type: string
        status:
          type: string
        model:
          type: string
        zone:
          type: string
        role:
          type: string
        origin:
          type: string
        activityScore:
          type: number
        activityWindow:
          type: string
        footprint:
          type: string
      required: [publicId, agentId, status, zone, role, origin]
    PublicStatus:
      type: object
      properties:
        connected:
          type: boolean
        status:
          type: string
        displayed:
          type: integer
        running:
          type: integer
        lastUpdatedAt:
          type: string
          format: date-time
      required: [connected, status, displayed, running, lastUpdatedAt]
...
    PublicTimelinePoint:
      type: object
      properties:
        capturedAt:
          type: string
          format: date-time
        connected:
          type: boolean
        status:
          type: string
        displayed:
          type: integer
        running:
          type: integer
        avgResponseTime:
          type: number
        toolCallCount:
          type: integer
        messageCount:
          type: integer
        zoneMix:
          type: array
          items:
            $ref: "#/components/schemas/PublicZoneMixEntry"
        modelMix:
          type: array
          items:
            $ref: "#/components/schemas/PublicModelMixEntry"
      required: [capturedAt, connected, status, displayed, running]
    PublicTimelineResponse:
      type: object
      properties:
        retentionHours:
          type: integer
        intervalSeconds:
          type: integer
        points:
          type: array
          items:
            $ref: "#/components/schemas/PublicTimelinePoint"
      required: [retentionHours, intervalSeconds, points]
    PublicReplayFrame:
      type: object
      properties:
        capturedAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/PublicStatus"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/PublicSession"
      required: [capturedAt, status, sessions]
    PublicReplayResponse:
      type: object
      properties:
        retentionHours:
          type: integer
        intervalSeconds:
          type: integer
        frames:
          type: array
          items:
            $ref: "#/components/schemas/PublicReplayFrame"
      required: [retentionHours, intervalSeconds, frames]
...
    PublicZonesHeatmapResponse:
      type: object
      properties:
        capturedAt:
          type: string
          format: date-time
        zones:
          type: array
          items:
            type: object
            properties:
              zone:
                type: string
              activityScore:
                type: number
              agentCount:
                type: integer
              statusDistribution:
                type: array
                items:
                  $ref: "#/components/schemas/PublicStatusShare"
            required: [zone, activityScore, agentCount, statusDistribution]
      required: [capturedAt, zones]
    PublicModelsDistributionResponse:
      type: object
      properties:
        models:
          type: array
          items:
            type: object
            properties:
              model:
                type: string
              share:
                type: number
              sampleCount:
                type: integer
              agentCount:
                type: integer
              avgResponseTime:
                type: number
              throughputTokensPerMinute:
                type: number
              avgLoad:
                type: number
            required: [model, share, sampleCount, agentCount, avgResponseTime, throughputTokensPerMinute, avgLoad]
      required: [models]
...
    PublicTrendPoint:
      type: object
      properties:
        capturedAt:
          type: string
          format: date-time
        onlineAgents:
          type: integer
        runningAgents:
          type: integer
        messageCount:
          type: integer
        totalTokens:
          type: integer
        deltaMessages:
          type: integer
        deltaToolCalls:
          type: integer
        deltaTokens:
          type: integer
        avgResponseTime:
          type: number
      required: [capturedAt, onlineAgents, runningAgents, messageCount, totalTokens, deltaMessages, deltaToolCalls, deltaTokens, avgResponseTime]
    PublicTrendSummary:
      type: object
      properties:
        sampleCount:
          type: integer
        onlineAgentsDelta:
          type: integer
        runningAgentsDelta:
          type: integer
        messageCountDelta:
          type: integer
        toolCallCount:
          type: integer
        totalTokensDelta:
          type: integer
        avgResponseTime:
          type: number
      required: [sampleCount, onlineAgentsDelta, runningAgentsDelta, messageCountDelta, toolCallCount, totalTokensDelta, avgResponseTime]
    PublicTrendResponse:
      type: object
      properties:
        hours:
          type: integer
        since:
          type: string
          format: date-time
        until:
          type: string
          format: date-time
        points:
          type: array
          items:
            $ref: "#/components/schemas/PublicTrendPoint"
        summary:
          $ref: "#/components/schemas/PublicTrendSummary"
      required: [hours, since, until, points, summary]
...
    PresenceEntry:
      type: object
      properties:
        deviceId:
          type: string
        online:
          type: boolean
      required: [deviceId, online]
    ChannelStatus:
      type: object
      properties:
        channel:
          type: string
        status:
          type: string
      required: [channel, status]
    NodeInfo:
      type: object
      properties:
        nodeId:
          type: string
        online:
          type: boolean
      required: [nodeId, online]
    CronJob:
      type: object
      properties:
        id:
          type: string
        schedule:
          type: string
        enabled:
          type: boolean
      required: [id, schedule, enabled]
...